estructura

+++ src/app/services/local-storage.service.ts
@@ -6,22 +6,36 @@ import { Injectable } from '@angular/core';
 export class LocalStorageService {
   // Método para guardar datos en `localStorage`
   setItem(key: string, value: any): void {
-    localStorage.setItem(key, JSON.stringify(value));
+    if (this.isLocalStorageAvailable()) {
+      localStorage.setItem(key, JSON.stringify(value));
+    }
   }
 
   // Método para obtener datos de `localStorage`
   getItem(key: string): any {
-    const item = localStorage.getItem(key);
-    return item ? JSON.parse(item) : null;
+    if (this.isLocalStorageAvailable()) {
+      const item = localStorage.getItem(key);
+      return item ? JSON.parse(item) : null;
+    }
+    return null;
   }
 
   // Método para eliminar un elemento de `localStorage`
   removeItem(key: string): void {
-    localStorage.removeItem(key);
+    if (this.isLocalStorageAvailable()) {
+      localStorage.removeItem(key);
+    }
   }
 
   // Método para limpiar todos los elementos de `localStorage`
   clear(): void {
-    localStorage.clear();
+    if (this.isLocalStorageAvailable()) {
+      localStorage.clear();
+    }
+  }
+
+  // Verificar si `localStorage` está disponible
+  private isLocalStorageAvailable(): boolean {
+    return typeof localStorage !== 'undefined';
   }
 }
